requirements.txt
tensorflow>=2.13.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
pandas>=2.0.0
pyserial>=3.5